requirements.txt
python-telegram-bot==22.0
python-dotenv==1.0.1
apscheduler==3.10.4
